uint attribute size reductions
by imposing the reasonable (?) limits explained in the README
we're able to store a lot of the structures using less space.
this should translate in lower memory usage, and possibly lower GC times
structures are also reorganized for better alingment in less space
Note that casting between different uint types seems to be very
efficient and worth the memory gains:
https://gist.github.com/Dieterbe/811e820e83ce05cac0a8
https://gist.github.com/Dieterbe/3f8894bf51daa97b3fc3

this requires more testing, especially for bugs and performance
regressions

+++ metric_tank/cwr.go
@@ -13,6 +13,6 @@ type ChunkReadRequest struct {
 type ChunkWriteRequest struct {
 	key       string
 	chunk     *Chunk
-	ttl       uint32
+	ttl       uint16
 	timestamp time.Time
 }
